move getAuthHeader to main
