+++ packages/insomnia/src/network/authentication.ts
@@ -1,145 +1,10 @@
-import * as Hawk from 'hawk';
-
-import type { AuthTypeOAuth2, RequestAuthentication, RequestParameter } from '~/insomnia-data';
+import type { RequestAuthentication, RequestParameter } from '~/insomnia-data';
 
 import type { RenderedRequest } from '../templating/types';
-import { COOKIE, HEADER, QUERY_PARAMS } from './api-key/constants';
-import { getBasicAuthHeader } from './basic-auth/get-header';
-import { getBearerAuthHeader } from './bearer-auth/get-header';
-import getOAuth1Token from './o-auth-1/get-token';
-import { getOAuth2Token } from './o-auth-2/get-token';
-
-interface Header {
-  name: string;
-  value: string;
-}
+import { QUERY_PARAMS } from './api-key/constants';
 
 export async function getAuthHeader(renderedRequest: RenderedRequest, url: string) {
-  const { method, body } = renderedRequest;
-  const authentication = renderedRequest.authentication as RequestAuthentication;
-
-  const requestId = renderedRequest._id;
-
-  if (authentication.disabled) {
-    return;
-  }
-
-  if (authentication.type === 'apikey' && authentication.addTo === HEADER) {
-    const { key, value } = authentication;
-    return {
-      name: key,
-      value: value,
-    } as Header;
-  }
-
-  if (authentication.type === 'apikey' && authentication.addTo === COOKIE) {
-    const { key, value } = authentication;
-    return {
-      name: 'Cookie',
-      value: `${key}=${value}`,
-    } as Header;
-  }
-
-  if (authentication.type === 'basic') {
-    const { username, password, useISO88591 } = authentication;
-    const encoding = useISO88591 ? 'latin1' : 'utf8';
-    return getBasicAuthHeader(username, password, encoding);
-  }
-
-  if (authentication.type === 'bearer' && authentication.token) {
-    const { token, prefix } = authentication;
-    return getBearerAuthHeader(token, prefix);
-  }
-
-  if (authentication.type === 'oauth2') {
-    // HACK: GraphQL requests use a child request to fetch the schema with an
-    // ID of "{{request_id}}.graphql". Here we are removing the .graphql suffix and
-    // pretending we are fetching a token for the original request. This makes sure
-    // the same tokens are used for schema fetching. See issue #835 on GitHub.
-    try {
-      const tokenId = requestId.match(/\.graphql$/) ? requestId.replace(/\.graphql$/, '') : requestId;
-      const oAuth2Token = await getOAuth2Token(tokenId, authentication as AuthTypeOAuth2);
-
-      if (oAuth2Token) {
-        const token = oAuth2Token.accessToken;
-        return _buildBearerHeader(token, authentication.tokenPrefix);
-      }
-      return;
-    } catch (err) {
-      // TODO: Show this error in the UI
-      console.log('[oauth2] Failed to get token', err);
-      return;
-    }
-  }
-
-  if (authentication.type === 'oauth1') {
-    const oAuth1Token = await getOAuth1Token(url, method, authentication, body);
-
-    if (oAuth1Token) {
-      return {
-        name: 'Authorization',
-        value: oAuth1Token.Authorization,
-      };
-    }
-    return;
-  }
-
-  if (authentication.type === 'hawk') {
-    const { id, key, algorithm, ext, validatePayload } = authentication;
-    let headerOptions = {
-      credentials: {
-        id,
-        key,
-        algorithm,
-      },
-      ext: ext,
-    };
-
-    if (validatePayload) {
-      const payloadValidationFields = {
-        payload: renderedRequest.body.text,
-        contentType: renderedRequest.body.mimeType,
-      };
-      headerOptions = Object.assign({}, payloadValidationFields, headerOptions);
-    }
-
-    const { header } = Hawk.client.header(url, method, headerOptions);
-    return {
-      name: 'Authorization',
-      value: header,
-    };
-  }
-
-  if (authentication.type === 'asap') {
-    const { issuer, subject, audience, keyId, additionalClaims, privateKey } = authentication;
-
-    let parsedAdditionalClaims;
-    try {
-      parsedAdditionalClaims = JSON.parse(additionalClaims || '{}');
-    } catch (err) {
-      throw new Error(`Unable to parse additional-claims: ${err}`);
-    }
-
-    if (parsedAdditionalClaims && typeof parsedAdditionalClaims !== 'object') {
-      throw new Error(`additional-claims must be an object received: '${typeof parsedAdditionalClaims}' instead`);
-    }
-    const generator = (await import('httplease-asap')).createAuthHeaderGenerator({
-      privateKey,
-      issuer,
-      keyId,
-      audience,
-      subject,
-      additionalClaims: parsedAdditionalClaims,
-      tokenExpiryMs: 10 * 60 * 1000, // Optional, max is 1 hour. This is how long the generated token stays valid.
-      tokenMaxAgeMs: 9 * 60 * 1000, // Optional, must be less than tokenExpiryMs. How long to cache the token.
-    });
-    return {
-      name: 'Authorization',
-      value: generator(),
-    };
-  }
-
-  return;
+  return (await import('../main/network/get-auth-header')).getAuthHeader(renderedRequest, url);
 }
 
 export function getAuthQueryParams(authentication: RequestAuthentication) {
